# packages/types/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "TechRat API",
    "description": "Gamified tech learning platform. Authenticate with POST /api/v1/auth/login (bearer) or ?useCookies=true (web).",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5099/"
    }
  ],
  "paths": {
...
    "/api/v1/practice/sessions": {
      "post": {
        "tags": [
          "Practice"
        ],
        "summary": "Start a practice session (Practice, Challenge, Random, Adaptive or a roadmap step)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartPracticeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PracticeSessionDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/practice/sessions/{id}": {
      "get": {
        "tags": [
          "Practice"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PracticeSessionDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/practice/sessions/{id}/answers": {
      "post": {
        "tags": [
          "Practice"
        ],
        "summary": "Grade an answer: returns feedback, explanation, reference, XP, levels, streak and roadmap progress",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitAnswerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnswerResultDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/analytics/me": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyticsDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/leaderboards/{scope}": {
      "get": {
        "tags": [
          "Leaderboard"
        ],
        "summary": "Global, Weekly, Monthly or Topic leaderboard",
        "parameters": [
          {
            "name": "scope",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/LeaderboardScope"
            }
          },
          {
            "name": "topic",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaderboardDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/search": {
      "get": {
        "tags": [
          "Search"
        ],
        "summary": "Search topics, subtopics and roadmaps",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SearchResultDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/notifications/{id}/read": {
      "post": {
        "tags": [
          "Notifications"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/v1/roadmaps": {
      "get": {
        "tags": [
          "Roadmaps"
        ],
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoadmapSummaryDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/roadmaps/{slug}": {
      "get": {
        "tags": [
          "Roadmaps"
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoadmapDetailDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/roadmaps/{slug}/start": {
      "post": {
        "tags": [
          "Roadmaps"
        ],
        "summary": "Enrol in a roadmap (fails with 403 while prerequisites are not met)",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoadmapDetailDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/topics/{slug}": {
      "get": {
        "tags": [
          "Topics"
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopicDetailDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/me": {
      "get": {
        "tags": [
          "Users"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSummaryDto"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Users"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSummaryDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/users/{username}/profile": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Public profile with topic levels, achievements, roadmap progress and activity",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileDto"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AnalyticsDto": {
        "required": [
          "days",
          "totalQuestions",
          "totalCorrect",
          "globalAccuracy",
          "studySeconds",
          "xpEarnedInPeriod",
          "currentStreak",
          "longestStreak",
          "xpLast7Days",
          "xpPrevious7Days",
          "velocityChangePercent",
          "byDifficulty",
          "byTopic",
          "bySubtopic",
          "daily",
          "strongestTopics",
          "weakestTopics"
        ],
        "type": "object",
        "properties": {
          "days": {
            "type": "integer",
            "format": "int32"
          },
          "totalQuestions": {
            "type": "integer",
            "format": "int32"
          },
          "totalCorrect": {
            "type": "integer",
            "format": "int32"
          },
          "globalAccuracy": {
            "type": "number",
            "format": "double"
          },
          "studySeconds": {
            "type": "integer",
            "format": "int32"
          },
          "xpEarnedInPeriod": {
            "type": "integer",
            "format": "int64"
          },
          "currentStreak": {
            "type": "integer",
            "format": "int32"
          },
          "longestStreak": {
            "type": "integer",
            "format": "int32"
          },
          "xpLast7Days": {
            "type": "number",
            "format": "double"
          },
          "xpPrevious7Days": {
            "type": "number",
            "format": "double"
          },
          "velocityChangePercent": {
            "type": "number",
            "format": "double"
          },
          "byDifficulty": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DifficultyBreakdownDto"
            }
          },
          "byTopic": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TopicAccuracyDto"
            }
          },
          "bySubtopic": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SubtopicAccuracyDto"
            }
          },
          "daily": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DailyActivityDto"
            }
          },
          "strongestTopics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TopicAccuracyDto"
            }
          },
          "weakestTopics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TopicAccuracyDto"
            }
          }
        }
      },
...
      "AnswerResultDto": {
        "required": [
          "feedback",
          "bonusXp",
          "totalXp",
          "level",
          "leveledUp",
          "topicSlug",
          "topicLevel",
          "topicXp",
          "topicLeveledUp",
          "currentStreak",
          "streakIncreased",
          "dailyChallengeBonusXp",
          "session",
          "completedSteps"
        ],
        "type": "object",
        "properties": {
          "feedback": {
            "$ref": "#/components/schemas/AnswerFeedbackDto"
          },
          "bonusXp": {
            "type": "integer",
            "format": "int32"
          },
          "totalXp": {
            "type": "integer",
            "format": "int64"
          },
          "level": {
            "$ref": "#/components/schemas/LevelDto"
          },
          "leveledUp": {
            "type": "boolean"
          },
          "topicSlug": {
            "type": "string"
          },
          "topicLevel": {
            "type": "integer",
            "format": "int32"
          },
          "topicXp": {
            "type": "integer",
            "format": "int64"
          },
          "topicLeveledUp": {
            "type": "boolean"
          },
          "currentStreak": {
            "type": "integer",
            "format": "int32"
          },
          "streakIncreased": {
            "type": "boolean"
          },
          "dailyChallengeBonusXp": {
            "type": "integer",
            "format": "int32"
          },
          "session": {
            "$ref": "#/components/schemas/SessionProgressDto"
          },
          "completedSteps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompletedStepDto"
            }
          }
        }
      },
...
      "LeaderboardDto": {
        "required": [
          "scope",
          "topicSlug",
          "periodStart",
          "page",
          "pageSize",
          "totalCount",
          "entries",
          "me"
        ],
        "type": "object",
        "properties": {
          "scope": {
            "$ref": "#/components/schemas/LeaderboardScope"
          },
          "topicSlug": {
            "type": [
              "null",
              "string"
            ]
          },
          "periodStart": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LeaderboardEntryDto"
            }
          },
          "me": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/LeaderboardEntryDto"
              }
            ]
          }
        }
      },
...
      "LeaderboardScope": {
        "enum": [
          "Global",
          "Weekly",
          "Monthly",
          "Topic"
        ]
      },
...
      "PracticeMode": {
        "enum": [
          "Practice",
          "Challenge",
          "Random",
          "Adaptive",
          "DailyChallenge",
          "Roadmap"
        ]
      },
      "PracticeSessionDto": {
        "required": [
          "id",
          "mode",
          "topicSlug",
          "topicName",
          "subtopicSlug",
          "subtopicName",
          "difficulty",
          "roadmapStepId",
          "isDailyChallenge",
          "totalQuestions",
          "answeredCount",
          "correctCount",
          "xpEarned",
          "startedAt",
          "completedAt",
          "questions"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "mode": {
            "$ref": "#/components/schemas/PracticeMode"
          },
          "topicSlug": {
            "type": [
              "null",
              "string"
            ]
          },
          "topicName": {
            "type": [
              "null",
              "string"
            ]
          },
          "subtopicSlug": {
            "type": [
              "null",
              "string"
            ]
          },
          "subtopicName": {
            "type": [
              "null",
              "string"
            ]
          },
          "difficulty": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Difficulty"
              }
            ]
          },
          "roadmapStepId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "isDailyChallenge": {
            "type": "boolean"
          },
          "totalQuestions": {
            "type": "integer",
            "format": "int32"
          },
          "answeredCount": {
            "type": "integer",
            "format": "int32"
          },
          "correctCount": {
            "type": "integer",
            "format": "int32"
          },
          "xpEarned": {
            "type": "integer",
            "format": "int32"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SessionQuestionDto"
            }
          }
        }
      },
      "ProfileDto": {
        "required": [
          "user",
          "topicProgress",
          "accuracyByDifficulty",
          "achievements",
          "roadmaps",
          "strongestTopics",
          "weakestTopics",
          "activity"
        ],
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserSummaryDto"
          },
          "topicProgress": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TopicProgressDto"
            }
          },
          "accuracyByDifficulty": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DifficultyBreakdownDto"
            }
          },
          "achievements": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AchievementDto"
            }
          },
          "roadmaps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoadmapProgressSummaryDto"
            }
          },
          "strongestTopics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TopicProgressDto"
            }
          },
          "weakestTopics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TopicProgressDto"
            }
          },
          "activity": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DailyActivityDto"
            }
          }
        }
      },
...
      "RoadmapDetailDto": {
        "required": [
          "summary",
          "prerequisites",
          "modules",
          "currentStepId"
        ],
        "type": "object",
        "properties": {
          "summary": {
            "$ref": "#/components/schemas/RoadmapSummaryDto"
          },
          "prerequisites": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoadmapPrerequisiteDto"
            }
          },
          "modules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoadmapModuleDto"
            }
          },
          "currentStepId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          }
        }
      },
...
      "RoadmapSummaryDto": {
        "required": [
          "id",
          "slug",
          "name",
          "description",
          "category",
          "difficulty",
          "estimatedHours",
          "stepsCount",
          "modulesCount",
          "icon",
          "xpReward",
          "prerequisites",
          "progress"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "slug": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "difficulty": {
            "type": "string"
          },
          "estimatedHours": {
            "type": "integer",
            "format": "int32"
          },
          "stepsCount": {
            "type": "integer",
            "format": "int32"
          },
          "modulesCount": {
            "type": "integer",
            "format": "int32"
          },
          "icon": {
            "type": "string"
          },
          "xpReward": {
            "type": "integer",
            "format": "int32"
          },
          "prerequisites": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "progress": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RoadmapUserStateDto"
              }
            ]
          }
        }
      },
...
      "SearchResultDto": {
        "required": [
          "type",
          "slug",
          "title",
          "subtitle",
          "icon",
          "url"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "subtitle": {
            "type": "string"
          },
          "icon": {
            "type": "string"
          },
          "url": {
            "type": "string"
          }
        }
      },
...
      "StartPracticeRequest": {
        "type": "object",
        "properties": {
          "mode": {
            "$ref": "#/components/schemas/PracticeMode"
          },
          "topicSlug": {
            "type": [
              "null",
              "string"
            ]
          },
          "subtopicSlug": {
            "type": [
              "null",
              "string"
            ]
          },
          "difficulty": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Difficulty"
              }
            ]
          },
          "count": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "roadmapStepId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          }
        }
      },
...
      "SubmitAnswerRequest": {
        "required": [
          "questionId",
          "selectedOptionId",
          "timeSpentSeconds"
        ],
        "type": "object",
        "properties": {
          "questionId": {
            "type": "string",
            "format": "uuid"
          },
          "selectedOptionId": {
            "type": "string",
            "format": "uuid"
          },
          "timeSpentSeconds": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "TopicDetailDto": {
        "required": [
          "topic",
          "progress",
          "subtopicProgress",
          "byDifficulty"
        ],
        "type": "object",
        "properties": {
          "topic": {
            "$ref": "#/components/schemas/TopicDto"
          },
          "progress": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TopicProgressDto"
              }
            ]
          },
          "subtopicProgress": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SubtopicProgressDto"
            }
          },
          "byDifficulty": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DifficultyBreakdownDto"
            }
          }
        }
      },
...
      "UpdateProfileRequest": {
        "required": [
          "displayName",
          "bio",
          "avatarUrl"
        ],
        "type": "object",
        "properties": {
          "displayName": {
            "type": [
              "null",
              "string"
            ]
          },
          "bio": {
            "type": [
              "null",
              "string"
            ]
          },
          "avatarUrl": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "UserSummaryDto": {
        "required": [
          "id",
          "username",
          "displayName",
          "email",
          "avatarUrl",
          "bio",
          "isAdmin",
          "level",
          "currentStreak",
          "longestStreak",
          "questionsAnswered",
          "correctAnswers",
          "accuracy",
          "globalRank",
          "createdAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "username": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "avatarUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "bio": {
            "type": [
              "null",
              "string"
            ]
          },
          "isAdmin": {
            "type": "boolean"
          },
          "level": {
            "$ref": "#/components/schemas/LevelDto"
          },
          "currentStreak": {
            "type": "integer",
            "format": "int32"
          },
          "longestStreak": {
            "type": "integer",
            "format": "int32"
          },
          "questionsAnswered": {
            "type": "integer",
            "format": "int32"
          },
          "correctAnswers": {
            "type": "integer",
            "format": "int32"
          },
          "accuracy": {
            "type": "number",
            "format": "double"
          },
          "globalRank": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  },
...
}